requirements.txt
Flask==2.2.2
python-decouple==3.6
wikipedia==1.4.0